Hello world

+++ A111/NUCLEO-L476RG_A111/Core/Inc/estimated_water_speed.h
@@ -4,9 +4,11 @@
  *  Created on: Oct 25, 2023
  *      Author: Bård Pedersen
  */
+#include <stdio.h>
+
 #ifndef ESTIMATED_WATER_SPEED_H
 #define ESTIMATED_WATER_SPEED_H
 
-double water_flow(double higth, double radius, double K, double S, double n, int number_peaks);
+uint16_t water_flow(double higth, double radius, double K, double S, double n, uint16_t number_peaks);
 
 #endif
